feito correções de tipos de dados

+++ src/main/java/br/com/techchallenge4/msPedidos/dto/ItemPedidoRequest.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ItemPedidoRequest {
     private Long idProduto;
-    private int quantidade;
+    private Long quantidade;
 }
